% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_funcs.R
\name{add_geneset_scores}
\alias{add_geneset_scores}
\title{Get proportion of counts for a cell assigned to a gene set}
\usage{
add_geneset_scores(object, geneset_list)
}
\arguments{
\item{object}{an SCESet object}

\item{geneset_list}{a GSEABase GeneSetCollection object}
}
\description{
Get proportion of counts for a cell assigned to a gene set
}
