% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_funcs.R
\name{df2sparse}
\alias{df2sparse}
\title{Function to convert dataframe of (gene,barcode,umi_count) to sparse Matrix}
\usage{
df2sparse(d)
}
\description{
Function to convert dataframe of (gene,barcode,umi_count) to sparse Matrix
}
