% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_funcs.R
\name{entropy}
\alias{entropy}
\title{Compute Shannon information entropy for a count vector}
\usage{
entropy(x)
}
\arguments{
\item{x}{a vector of counts}
}
\value{
the entropy of the corresponding multinomial distribution
}
\description{
Compute Shannon information entropy for a count vector
}
