% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_funcs.R
\name{get_barcode_stats}
\alias{get_barcode_stats}
\title{Get barcode stats}
\usage{
get_barcode_stats(mat)
}
\arguments{
\item{mat}{numeric matrix, rows represent genes, columns represent cells}
}
\value{
a data_frame with barcode-level statistics
}
\description{
Get barcode stats
}
