% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_funcs.R
\name{get_gene_stats}
\alias{get_gene_stats}
\title{Get gene-level stats}
\usage{
get_gene_stats(mat)
}
\arguments{
\item{mat}{numeric matrix, rows represent genes, columns represent cells}
}
\value{
a data_frame with gene-level statistic
}
\description{
Get gene-level stats
}
