% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_funcs.R
\name{get_geneset_score}
\alias{get_geneset_score}
\title{Get proportion of counts for a cell assigned to a gene set}
\usage{
get_geneset_score(object, geneset)
}
\arguments{
\item{object}{an SCESet object}

\item{geneset}{character vector of HGNC symbols for the given gene set}
}
\description{
Get proportion of counts for a cell assigned to a gene set
}
