% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_funcs.R
\name{profile_ambient}
\alias{profile_ambient}
\title{Profile ambient genes}
\usage{
profile_ambient(object, max_umis = 1000)
}
\arguments{
\item{object}{an SCESet object}

\item{max_umis}{numeric, maximum number of total UMIs to define a "low total"
barcode}
}
\value{
list with aggregated counts and aggregated counts-per-million
}
\description{
Use barcodes with low total UMI counts to obtain aggregate profile
of gene expression.
}
